+++ DesktopApplicationTemplate.UI/Views/MqttCreateServiceView.xaml.cs
@@ -1,13 +1,16 @@
 using System.Windows.Controls;
+using DesktopApplicationTemplate.Core.Services;
+
 using DesktopApplicationTemplate.UI.ViewModels;
 
 namespace DesktopApplicationTemplate.UI.Views;
 
 public partial class MqttCreateServiceView : Page
 {
-    public MqttCreateServiceView(MqttCreateServiceViewModel viewModel)
+    public MqttCreateServiceView(MqttServiceViewModel vm, ILoggingService logger)
     {
         InitializeComponent();
-        DataContext = viewModel;
+        DataContext = vm;
+        vm.Logger = logger;
     }
 }